created validators for each route of both services

+++ src/controllers/recipe.controller.js
@@ -1,4 +1,6 @@
-const {recipeModel} = require("../models/recipe.models")
+const {recipeModel} = require("../models/recipe.models");
+const validators = require("../validators/recipe.validator");
+const {formatZodError} = require("../utilities/errormessage")
 
 // get all recipes
 async function getAllRecipes(req, res) {
@@ -9,12 +11,24 @@ async function getAllRecipes(req, res) {
 
 // get single recipe
 async function getSingleRecipe(req, res) {
+   const result = validators.getSingleRecipeValidator.safeParse(req.body)
+
+   if (!result.success) {
+      return res.status(400).json(formatZodError(result.error.issues)).end();
+   }
+
     const recipe = await recipeModel.findById(req.params.recipeId)
     res.json(recipe).end();
 }
 
 // add new recipe
 async function addRecipe(req, res) {
+   const result = validators.recipeValidator.safeParse(req.body)
+
+   if (!result.success) {
+      return res.status(400).json(formatZodError(result.error.issues)).end();
+   }
+
     await recipeModel.create  ({
       name: req.body.name,
       process: req.body.process,
@@ -25,6 +39,12 @@ async function addRecipe(req, res) {
 
 // update recipe
 async function udpateRecipe(req, res) {
+   const result = validators.updateRecipesValidator.safeParse(req.body)
+
+   if (!result.success) {
+      return res.status(400).json(formatZodError(result.error.issues)).end();
+   }
+
     await recipeModel.updateOne({_id: req.params.recipeId}, {...req.body});
 
     res.send("recipe updated successfully!").end();
@@ -32,6 +52,12 @@ async function udpateRecipe(req, res) {
 
 // delete recipe
 async function deleteRecipe(req, res) {
+   const result = validators.deleteRecipeValidator.safeParse(req.body)
+
+   if (!result.success) {
+      return res.status(400).json(formatZodError(result.error.issues)).end();
+   }
+
     await recipeModel.deleteOne({_id: req.params.recipeId});
     res.send("recipe deleted").end();
 }
